Refactor ECPrivateKey.freshPrivateKey to use CryptoUtil.freshPrivateKey
